Solucionado problema de division entre y se añadieron bugs

+++ src/proyectofinal/Cliente.java
@@ -63,7 +63,7 @@ public class Cliente {
         this.tiempoDePrestamo = tiempoDePrestamo;
     }
 
-    public float getTotalPrestado() {
+    public int getTotalPrestado() {
         return totalPrestado;
     }
